Replaced deprecated `dragon-code/simple-dto` with `spatie/laravel-data`
