+++ composer.json
@@ -40,13 +40,13 @@
     "require": {
         "php": "^8.2",
         "composer-runtime-api": "^2.2",
-        "dragon-code/simple-dto": "^2.5.1",
         "dragon-code/support": "^6.6",
         "illuminate/console": "^10.0 || ^11.0 || ^12.0",
         "illuminate/container": "^10.0 || ^11.0 || ^12.0",
         "illuminate/database": "^10.0 || ^11.0 || ^12.0",
         "illuminate/support": "^10.0 || ^11.0 || ^12.0",
         "laravel/prompts": ">=0.1",
+        "spatie/laravel-data": "^4.14",
         "symfony/console": "^6.0 || ^7.0"
     },
     "require-dev": {
